requirements.txt
PyQt5==5.15.11
pikepdf==4.4.0
Pillow==8.2.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pdf2image==1.17.0
img2pdf==0.5.1
pillow==10.4.0
PyPDF2==3.0.1